Manacher's Algo
Manacher's algorithm is used to find the longest palindromic substring in any string. It is 
required to solve sub-problems of some very hard problems.
Manacher's algorithm was invented by Manacher for listing all the palindromes that appear at 
the start of any given string, it was later observed that the same algorithm can be used to 
find the longest palindromic substring of any string in linear time.